packages/message_broker/requirements.txt
aio-pika==9.4.3
aiodns==3.2.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiormq==6.8.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
pamqp==3.3.0
pika==1.3.2
propcache==0.2.0
pycares==4.4.0
pycparser==2.22
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.16.0
